Removed unnecessary develop files
