--- functions/rgb_colors.py
@@ -1,5 +1,0 @@
-
-
-red = (255,0,0)
-green = (0,255,0)
-blue = (0,0,255)
