+++ maragoniProcessing/videoHandling.py
@@ -1,39 +1,46 @@
+import tkinter
+from tkinter.filedialog import askopenfile
+from pathlib import Path
 from tqdm import tqdm
+
 import cv2
 import numpy as np
-from functions.getFile import fileNameHandler
- 
 
-class videoWriterClass:
+
+class videoHandler():
     def __init__(self):
         '''Opens the video with the help of a file handler'''
-        self.fHandler = fileNameHandler()
+        self.fHandler = filenameHandler()
         self.vidCapIn = cv2.VideoCapture(self.fHandler.inFilename)
 
         if not self.vidCapIn.isOpened():
            print('Cannot open video - Script aborted')
            exit(1)
+
+        width = int(self.vidCapIn.get(cv2.CAP_PROP_FRAME_WIDTH))
+        height = int(self.vidCapIn.get(cv2.CAP_PROP_FRAME_HEIGHT))
+        self.vidDim = (height, width)
             
     def creatOutputVideo(self, fcn: callable ,vidDim: tuple) -> None:
         '''creates an output video by a given warping function'''
 
-        # creates the output video handler
-        fourc = int(self.vidCapIn.get(cv2.CAP_PROP_FOURCC))
-        fps = int(self.vidCapIn.get(cv2.CAP_PROP_FPS))
         if vidDim is None:
-            # new video has same size as old video
-            width = int(self.vidCapIn.get(cv2.CAP_PROP_FRAME_WIDTH))
-            height = int(self.vidCapIn.get(cv2.CAP_PROP_FRAME_HEIGHT))
-            vidDim = (width, height)
+            vidDim = self.vidDim
+
+        # creates the output video handler
+        fourcc = int(self.vidCapIn.get(cv2.CAP_PROP_FOURCC))
+        #fourcc = cv2.VideoWriter_fourcc(*'')
+        fps = int(self.vidCapIn.get(cv2.CAP_PROP_FPS))
         self.Nframes = int(self.vidCapIn.get(cv2.CAP_PROP_FRAME_COUNT))
 
         try:
-            self.vidCapOut = cv2.VideoWriter(self.fHandler.outFilename, fourc, fps, vidDim)
+            self.vidCapOut = cv2.VideoWriter(self.fHandler.outFilename, fourcc, fps, vidDim)
         except AttributeError:
             print("Add suffix to create an Ouptu Videowriter!")
         
         # iterate over video
         self.vidCapIn.set(cv2.CAP_PROP_POS_FRAMES, 0) # start at the beginning
+        print('Start creating Video')
         pbar = tqdm(total=self.Nframes)
         while(self.vidCapIn.isOpened()):
             ret, frame = self.vidCapIn.read() # Capture frame-by-frame
@@ -42,13 +49,19 @@ class videoWriterClass:
                 self.vidCapOut.write(frame)
                 pbar.update(1) # update progress bar            
             else: # Break the loop
-                print('cannot open video')
+                pbar.close()
+                print('VideoWriter closed')
                 break     
         self.vidCapIn.release(), self.vidCapOut.release()
+    
+    @property
+    def dimension(self):
+        return self.vidDim
 
             
     
     def getSingleImage(self, frameNumber: int) -> np.ndarray:
+        '''returns a frame selected by a frame number'''
         self.vidCapIn.set(cv2.CAP_PROP_POS_FRAMES, frameNumber)
         ret, img = self.vidCapIn.read()
         if ret == True:
@@ -65,27 +78,38 @@ class videoWriterClass:
         yield StopIteration
 
 
-#from functions.diameters import burstingDiameter, spreadDiameter
-import seaborn as sns
-import matplotlib as plt
 
+
+class filenameHandler():
+    def __init__(self) -> None:
+        # open file dialog
+        root = tkinter.Tk()
+        root.withdraw()
+
+        fname = askopenfile(title="Select Videofile")
+        self.filename = Path(fname.name)
+
+    @property
+    def inFilename(self):
+        return str(self.filename.absolute())
+
+    @property
+    def outFilename(self):
+        fname = self.filename
+        output = fname.with_name(f"{fname.stem}_out{fname.suffix}")
+        return str(output.absolute())
+
+
+
+## TEST my CLASS
 if __name__ == '__main__':
-    def myPolar(frame: np.ndarray) -> np.ndarray:
-        
-        (h,w) = frame.shape[0:2]
-        center = (int(h/2), int(w/2))
-        r = int(h/2)
-        polar_image = cv2.warpPolar(
-            src=frame,
-            center=center,
-            dsize=(w,h),
-            maxRadius=r, 
-            flags=cv2.WARP_FILL_OUTLIERS)
-        return polar_image
+    myHandler = filenameHandler()
+    print(myHandler.inFilename)
+    print(myHandler.outFilename)
 
-
-    myWriter = videoWriterClass()
+  
+    myWriter = videoHandler()
     img = myWriter.getSingleImage(frameNumber=0)
     (w,h) = img.shape[0:2]
-    r = int(h/2)
-    myWriter.creatOutputVideo(vidDim = (w,h), fcn = myPolar)
+    print(w)
+    print(h)
